# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CommentRequest:
    properties:
      task_id:
        type: integer
      text:
        type: string
    type: object
  dto.CommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      task_id:
        type: integer
      text:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.ProjectRequest:
    properties:
      description:
        type: string
      member_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      status:
        type: string
    type: object
  dto.ProjectResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      manager_id:
        type: integer
      member_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API для системы управления проектами и задачами с поддержкой навыков
    сотрудников
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SkillsTracker API
  version: 1.0.0
paths:
  /api/v1/comments:
    get:
      description: Возвращает все комментарии к указанной задаче, отсортированные
        по дате создания.
      parameters:
      - description: ID задачи
        in: query
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить комментарии по task_id
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Создает новый комментарий к задаче. Автором комментария становится
        текущий авторизованный пользователь.
      parameters:
      - description: Данные комментария (task_id, text)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать комментарий
      tags:
      - comments
  /api/v1/comments/{id}:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить комментарий
      tags:
      - comments
    get:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить комментарий по id
      tags:
      - comments
    put:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные комментария
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить комментарий
      tags:
      - comments
...
  /api/v1/projects:
    get:
      description: Получить список проектов. Если указан manager_id, возвращаются
        проекты конкретного менеджера. Иначе возвращаются все проекты.
      parameters:
      - description: ID менеджера (для фильтрации по менеджеру)
        in: query
        name: manager_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProjectResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить проекты
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Создает новый проект. Только менеджеры могут создавать проекты.
        Создатель автоматически становится менеджером проекта. Можно сразу добавить
        участников проекта.
      parameters:
      - description: Данные проекта (name, description, status, member_ids)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать проект
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить проект
      tags:
      - projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить проект по id
      tags:
      - projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные проекта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить проект
      tags:
      - projects
  /api/v1/projects/{id}/members:
    delete:
      description: Удаляет пользователя из проекта. Менеджера проекта удалить нельзя.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID (ID пользователя для удаления)
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить участника из проекта
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Добавляет пользователя в проект. Только менеджер проекта может
        добавлять участников.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID (ID пользователя для добавления)
        in: query
        name: user_id
        required: true
        type: integer
      - description: 'Роль в проекте (по умолчанию: developer)'
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить участника в проект
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"